Testeo Dijkstra

+++ src/main/java/Start.java
@@ -43,7 +43,7 @@ public class Start {
 
     // voy agregando los nodos
     for (CSVRecord record : records) {
-      graph.addNode(record.get("stop_id"),record.get("route_short_name"), Float.parseFloat(record.get("stop_lat")), Float.parseFloat(record.get("stop_lon")), Integer.parseInt(record.get("direction_id")));
+      graph.addNode(record.get("stop_id"),record.get("route_short_name"), Double.parseDouble(record.get("stop_lat")), Double.parseDouble(record.get("stop_lon")), Integer.parseInt(record.get("direction_id")));
     }
     graph.addEdges();
     System.out.println(graph.getSize());
